typeorm migration(comments and likes)

+++ src/domains/users/use-cases/delete-user-use-case.ts
@@ -1,5 +1,6 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { UsersRepository } from '../../../infrastructure/repositories/users.repository';
+import { isUUID } from '../../../utils/utils';
 
 export class DeleteUserCommand {
   constructor(public id: string) {}
@@ -10,6 +11,7 @@ export class DeleteUserUseCase implements ICommandHandler<DeleteUserCommand> {
   constructor(private readonly usersRepository: UsersRepository) {}
 
   async execute(command: DeleteUserCommand): Promise<boolean> {
+    if (!isUUID(command.id)) return null;
     return await this.usersRepository.deleteUser(command.id);
   }
 }
